+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Removed
+
+- Removed unused import
+
+### Fixed
+
+- Fixed typo in varialbe
 
 ## [0.1.0] - 2018-05-12